% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleData.R
\docType{data}
\name{cachexiaData}
\alias{cachexiaData}
\title{Metabolite concentrations}
\format{A data frame with 77 rows and 65 variables:}
\source{
\url{http://dx.doi.org/10.1007/s11306-010-0232-9}
}
\usage{
data(cachexiaData)
}
\description{
A dataset containing the metabolite concentrations of 77 urine samples 
from cancer patients measured by 1H NMR
}
\keyword{datasets}

