% Generated by roxygen2 (4.0.2): do not edit by hand
\name{forestPerformance}
\alias{forestPerformance}
\title{Characterizing the performance of a Random Forest model}
\usage{
forestPerformance(cm)
}
\arguments{
\item{cm}{the confusion matrix of a validated Random Forest model}
}
\value{
a vector contining the accuracy, the recall and the same confusion matrix of the Random Forest model
}
\description{
This function provides the accuracy and the recall of a Random Forest model
}
\author{
Piergiorgio Palla
}

