% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFMCcv.R
\name{getAvgAUC}
\alias{getAvgAUC}
\title{Computing the average AUC}
\usage{
getAvgAUC(predictions, labels)
}
\arguments{
\item{predictions}{a vector, matrix, list, or data frame containing the predictions.}

\item{labels}{a vector, matrix, list, or data frame containing the true class labels. It must have
the same dimensions as predictions}
}
\value{
the average AUC value
}
\description{
This function computes the average AUC of the ROC curves obtained from a cross-validation procedure
}
\examples{
## load a simple dataset with the vectors of the predictions and the labels resulting from a CV 
data(simpleData)
avg_auc <- getAvgAUC(simpleData$predictions, simpleData$labels)
}
\author{
Piergiorgio Palla
}

