% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploring.R
\name{mds}
\alias{mds}
\title{mds class}
\usage{
mds(
  expr_df,
  opt = list(ntree = 1000, mtry = round(sqrt(ncol(expr_df) - 2)), seed = 1)
)
}
\arguments{
\item{expr_df}{a n x p dataframe representing the expr_df to be used to train the model. The first two columns
must represent respectively the samples names and the class labels related to each sample}

\item{opt}{a list of optional parameters useful to train the random forest. It may include the number
of trees (ntree), the parameter mtry and the seed}
}
\value{
an object of class mds including the following attributes: \itemize{
\item model an object of class random forest containing the proximity matrix
\item classes a factor with the classes associated with each sample, used to train the model
\item sample_names the vector of the names of the samples
}
}
\description{
A constructor function for the S3 class mds; the mds class encapsulates useful information for generating MDS plots
}
\examples{
data(cachexiaData)
params <- list(ntree = 50, mtry = round(sqrt(ncol(cachexiaData[,1:10]) - 2)), seed = 1)
mds_obj <- mds(cachexiaData[,1:10], opt = params)

}
\author{
Piergiorgio Palla
}
