% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploring.R
\name{pca.loadings.plot}
\alias{pca.loadings.plot}
\title{PCA Loadings plot}
\usage{
pca.loadings.plot(pca_obj, nvar)
}
\arguments{
\item{pca_obj}{an object of class pca}

\item{nvar}{the number of variables to plot}
}
\description{
This function plots the relation between the original variables and the subspace dimensions.
It is useful for interpreting relationships among variables.
}
\examples{
data(cachexiaData)
pca_obj <- pca(cachexiaData[,1:10], autoscale = TRUE, exclude = TRUE)
pca.loadings.plot(pca_obj, nvar = 20)

}
\author{
Piergiorgio Palla
}
