% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploring.R
\name{pca.scores.plot}
\alias{pca.scores.plot}
\title{PCA Scores plot}
\usage{
pca.scores.plot(pca_obj, expr_df, xrange, yrange)
}
\arguments{
\item{pca_obj}{an object of class pca}

\item{expr_df}{a n x p dataframe representing the expr_df used to create the pca object}

\item{xrange}{a vector of two elements indicating the range of values along the x-axis in which
eventually it will be specified the name of the samples}

\item{yrange}{a vector of two elements indicating the range of values along the y-axis in which
eventually it will be specified the name of the samples}
}
\description{
This function creates a plot that graphically projects the original samples onto the subspce
spanned by the first two principal components
}
\details{
The Scores plot is used for interpreting relations among the observations
}
\examples{
data(cachexiaData)
pca_obj <- pca(cachexiaData[,1:10], autoscale = TRUE, exclude = TRUE)
pca.scores.plot(pca_obj, cachexiaData[,1:10])

}
\author{
Piergiorgio Palla
}
