% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exploring.R
\name{plot.pca.scores}
\alias{plot.pca.scores}
\title{PCA Scores plot
  
This function creates a plot that graphically projects the original samples onto the subspce
spanned by the first two principal components}
\usage{
\method{plot}{pca.scores}(pca_obj, dataset, xrange, yrange)
}
\arguments{
\item{pca_obj}{an object of class pca}

\item{dataset}{a n x p dataframe representing the dataset used to create the pca object}

\item{xrange}{a vector of two elements indicating the range of values along the x-axis in which
eventually it will be specified the name of the samples}

\item{yrange}{a vector of two elements indicating the range of values along the y-axis in which
eventually it will be specified the name of the samples}
}
\description{
PCA Scores plot
  
This function creates a plot that graphically projects the original samples onto the subspce
spanned by the first two principal components
}
\details{
The Scores plot is used for interpreting relations among the observations
}
\examples{
## data(cachexiaData)
## pca_obj <- pca(cachexiaData, autoscale = TRUE, exclude = TRUE) 
## plot.pca.scores(pca_obj, cachexiaData)
}
\author{
Piergiorgio Palla
}

