% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exploring.R
\name{screeplot}
\alias{screeplot}
\title{Scree Plot}
\usage{
screeplot(pca_obj, ncomp)
}
\arguments{
\item{pca_obj}{an object of class pca}

\item{ncomp}{the number of components to plot}
}
\description{
This function creates  a graphical display of the variances against the number of principal components.
It is used to determine how many components should be retained in order to explain a high percentage of the variation in the data
}
\details{
screplot generates 2 graphs that represent respectively the relative variances
and the cumulative variances associated with the principal components
}
\examples{
## data(cachexiaData)
## pca_obj <- pca(cachexiaData, autoscale = TRUE, exclude = TRUE)
## screeplot(pca_obj, ncomp = 10)
}
\author{
Piergiorgio Palla
}

