% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{simpleData}
\alias{simpleData}
\title{simpleData}
\format{\preformatted{List of 2
 $ predictions: num [1:200] 0.868 0.544 0.212 0.627 0.432 ...
 $ labels     : num [1:200] 0 0 0 0 0 0 0 0 0 0 ...
}}
\usage{
data(simpleData)
}
\description{
A list of two vectors named respectively predictions and labels
}
\keyword{datasets}

