% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleData.R
\docType{data}
\name{simpleData}
\alias{simpleData}
\title{simpleData}
\format{An object of class \code{list} of length 2.}
\usage{
data(simpleData)
}
\description{
A list of two vectors named respectively predictions and labels
}
\keyword{datasets}

